requirements.txt
django==2.0.6
psycopg2-binary==2.7.5
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
django-graphql-jwt==0.1.10
graphene==2.1.2
graphene-django==2.1rc1
snapshottest==0.5.0
rescape-graphene==0.1.18
numpy==1.14.5
django-graphql-geojson==0.1.4
